pyproject.toml
[project]
name = "whispera"
version = "1.0.0"
description = "Transcribe audio recordings to text using OpenAI Whisper API"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyinstaller>=6.0.0",
    "pydub>=0.25.1",
]
